Fix compositor GPU usage calculation during drops
CPM frame drops sometimes come through as 0 GPU duration. Previously we
were using the duration of the perfetto event in that case, but that is
also way too big in some cases so it can completely skew the results.
For example if the compositor is dropping lots of frames because of
non-preemptable contention the results were off.

This computes an average of all compositor GPU slice durations and uses
that whenever it encounters a 0 result. Since the compositor is mostly
consistent from frame to frame this produces much better results in the
face of some missing data.

Bug: 470155455
Fixed: 470155455
Change-Id: I207e8c7af6aaedd3fd89707f4764834ca54004a1